refactor: Remove unnecessary type annotations from tests
- Remove redundant `as int` casts on integer literals (42 is already int)
- Remove explicit generic type parameters when inferable from arguments
- Simplify lambda parameter/return types when contextually typed
- Keep necessary narrowing casts (as byte, as short, as long, as float)
- Keep `as long` for values exceeding int32 range

Updated 43 test files (E2E fixtures and golden tests) to use
idiomatic Tsonic patterns that rely on type inference.

+++ packages/emitter/testcases/common/classes/generic-inheritance/ConcreteExtends.ts
@@ -14,12 +14,12 @@ export class Container<T> {
 // Non-generic extending generic with concrete type
 export class IntContainer extends Container<int> {
   double(): int {
-    return (this.getValue() * 2) as int;
+    return this.getValue() * 2;
   }
 }
 
 export class StringContainer extends Container<string> {
   getLength(): int {
-    return this.getValue().length as int;
+    return this.getValue().length;
   }
 }
