+++ test/fixtures/nullish-coalescing-threading/src/index.ts
@@ -20,9 +20,9 @@ function getFirstValid(a: int | null, b: int | null): int {
 
 // Test 3: ?? in return with arithmetic RHS
 function divideOrDefault(value: int | null, divisor: int): int {
-  const a = 10 as int;
-  const b = 3 as int;
-  return value ?? ((a / b) as int); // Should use int division (3, not 3.33)
+  const a = 10;
+  const b = 3;
+  return value ?? a / b; // Should use int division (3, not 3.33)
 }
 
 export function main(): void {
@@ -31,21 +31,21 @@ export function main(): void {
 
   // Test 1
   const r1 = getOrDefault(null);
-  const r2 = getOrDefault(77 as int);
+  const r2 = getOrDefault(77);
   Console.writeLine(`getOrDefault(null) = ${r1}`); // 100
   Console.writeLine(`getOrDefault(77) = ${r2}`); // 77
 
   // Test 2
   const r3 = getFirstValid(null, null);
-  const r4 = getFirstValid(null, 50 as int);
-  const r5 = getFirstValid(25 as int, 50 as int);
+  const r4 = getFirstValid(null, 50);
+  const r5 = getFirstValid(25, 50);
   Console.writeLine(`getFirstValid(null, null) = ${r3}`); // 333
   Console.writeLine(`getFirstValid(null, 50) = ${r4}`); // 50
   Console.writeLine(`getFirstValid(25, 50) = ${r5}`); // 25
 
   // Test 3
-  const r6 = divideOrDefault(null, 3 as int);
-  const r7 = divideOrDefault(99 as int, 3 as int);
+  const r6 = divideOrDefault(null, 3);
+  const r7 = divideOrDefault(99, 3);
   Console.writeLine(`divideOrDefault(null, 3) = ${r6}`); // 3 (int division)
   Console.writeLine(`divideOrDefault(99, 3) = ${r7}`); // 99
